requirements.txt
Flask==1.1.1
Flask-RESTful==0.3.7
Flask-HTTPAuth==3.3.0
Markdown==3.1.1
mysql-connector==2.2.9